backend/requirements.txt
flask===2.2.2
Flask-Cors===3.0.10
click==8.1.3
numpy==1.23.3
pandas==1.5.0
python-dateutil==2.8.2
gunicorn==20.1.0
python-dotenv==0.21.0
pytz==2022.5
PyGithub==1.56
